Updated view model to support observing activities change
